third_party/blink/web_tests/platform/linux-chrome/external/wpt/fetch/stale-while-revalidate/stale-image-expected.txt

This is a testharness.js-based test.
[FAIL] Cache returns stale resource
  assert_equals: (revalidated version loaded) expected 256 but got 16
Harness: the test ran to completion.
